style(folder): move user repo and tests to user folder
